switches to CPS monad for the Machine
Apparently, just using the state monad is not enough, as in that case,
we can switch between tasks only in some predefined places (like sync
points). An attempt to fork in an abitrary point will lead to lost
computation after a resume.

Thus, the only solution is to use the Continutation monad. This commit
makes the Machine monad a continuation monad, except that as I'm a
greedy bastard, I didn't make call/cc public -- it is just used to
implement fork/switch of the multi-state monad interface.

The new monad layout is the following F.C.S.M, where
   - F is a failure monad
   - C is a continuation monad
   - S is a state monad
   - M is the inner monad.

It is also easy to remember - Football Club Spartak Moscow

The new behavior of a computation in the Machine monad is consolidated
in the fork and switch methods.

The `fork` function will remember current continuation in the current
state, and then switch and restore the continuation of the new state,
e.g., resume the computation, that was previously saved. The `switch`
will basically do the same, except that instead of creating a new
state, it will switch to the existing one.