reimplements the interpreter in the Cont monad
Since we reified the computation, we don't need the next pointer any
more. The new implementation is much easier to comprehend and relies
on a linker - we implement computations for blk and sub, and link them
into a program via the linker interface, so that there is only one
component who is responsible for code finding.

Also, the linker no longer depends on the interpreter, so the code
type is simplified - it is just a functor that returns a computation
in the given monad.