moves Primus to the new monad
All code is finally updated, though the system shouldn't work at this
time, as there are few stubs here and there.

We are moving away from distinguishing a succesfull or unsuccessful
computation, but instead we are denoting a result of a computation to
be either normal, or exceptional. In both cases we actually have a
result.