fixes multiple bugs in the machine implementation
The main bug, that was preventing a machine from switching to a clone
after an observation, is very easy, I just myself, prevented it, by
wrapping the provide call with the `with_global` context handler, that
was always restoring the state to global.

The rest bugs are minor issues.

So far, it looks like that the non-deterministic machine works as expected.