reduces the number of forks in a promiscuous mode
For each basic block a computation is forked only once.

A question here, is whether we should decide how much to fork in the
execution mode, or in the scheduler. Or maybe we should add some other
component, that will be responsible for reducing a number of forks.

The question is let open, so a decision to make promiscous mode of
execution to fork only once is arbitrary and may/will be revised in
future.