new bitvector representation (#680)
The proposed representation uses the Z.t as a container for both: the
payload, and the metadata.

We reserve 15 bits for the metadata. The first bit represents whether a
bitvector should be treated as signed or unsigned. The rest 14 bits
contain the size of a bitvector. With such representation we can encode
only bitvectors of size 2^14 or less. This is, however, enough in
practice. If more would be needed we may switch to a more complex
prefixed encoding akin to UTF.

```
     +-----------+------+---+
     |  payload  | size | s |
     +-----------+------+---+
      size+15  15 14       0
```

Given this representation all bitvectors that are smaller than
0x1_000_000_000_000 are represented as an OCaml int, and are thus
unboxed, so creating new bitvector, and performing bitvector
arithmentics no longer perfoms allocation.

```
 # let deadbeef = Word.of_int32 0xDEADBEEFl;;
 val deadbeef : Bap.Std.Word.t = 0xdeadbeef:32
 # Obj.is_int (Obj.repr deadbeef);;
 - : bool = true
```

The previous representation is still available under
`Word.Stable.V1`. It is useful for serialization and deserialization of
data stored in previous representation. In particular, we are using the
old representation to read byteweight signatures.