reduces the memory representation RAM consumption (#682)
As a continuation of #680, this PR will reduce the memory footprint of
the memory representation in bap, the `mem` type. As an optimization, we
were precomputing a memory accessors during the construction of a memory
object. Although, it might save some time and didn't at all, as the
memory accessor operations are rarely used if used at all. Thus, this PR
removes the precomputed getters from the memory, and moves all the check
to the time of getter invokation.

As a result, this reduces the size of memory representation by 10
words (80 bytes on amd64). In total with the #680, this will reduce
memory representation from 21 words to only 7.