implements new Bitvector pretty printer
and rolls back the [Word.string_of_value] behavior.

The pretty printing function was totally rewritten from scratch, as
Z.format and other Z printing functions are broken. Now we have a
generic printer that suits all tastes, as well as 9 preinstantiated
pretty printers.

We also changed the bitvector default string representation. The new
representation doesn't used [true] or [false] for the [1] and [0] (this
is binary analysis anyway), and doesn't loose the signedness
information.