automates documentation generation
Now it is enough just to type `make doc` and everything will work out of
box, if a correct version of OCaml is used, i.e., 4.03. The Makefile
recipe will pull the latest version of argot (git is required) and use
it to build the documentation. BAP should be installed either from the
source tree, or from opam, doesn't matter.