makes it compilable with the new core
Most of the changes are due to the following breaking changes in the
upstream:

1. change of the monad interface, now `bind` takes a function as a
labeled argument. This is the most annoying change, as we need to update
our Base interface to be consistent with Core, but this will force us to
release the 2.0 version.

2. The `equal` parameter for different finders now became mandatory (to
discourage people from using the polymorphic compare).

This will fix #648.