supports building without ppx_deriving
the ppx_deriving is an optional depedency of many ppx_X packages from
Janestreet. The optionality is specified using a predicated
requirement. The problem is mostly in ocamlfind, that if there is no
`-pp` option will perform a query is made to build a preprocessor
command that doesn't include the predicates passed through the command
line, that lead to a failure, since ppx_deriving is not installed.

The workaround is simple - we need to pass the `-pp` option to the
linking phase - kind of stupid, but it will prevent the ocamlfind driver
from making this query.

The bap_build library was also update, I've added a set of default
predicates that are used when the set of bap dependencies is inferred.