disables the bap_elf library
So due to issues with [ocamlfind][1] and the new generation of the ppx derivers
from the Janestreet we can't build the bap_elf library without a bogus
dependecy on ppx_deriving. A trick to supress this dependency from
jumping in was to specify a `-pp` flag, and we can't use this trick with
a library that uses camlp4. So we can only build it, if ppx_deriving is
installed (it is not actually used, it's needed just to keep ocamlfind
happy).

Since `ppx_deriving` is not really needed, and it usually lags behind
the compiler development, we decided to disable this library by
default, until we will rewrite it using `ppx_bitstrings`.

[1]: https://gitlab.camlcity.org/gerd/lib-findlib/issues/19