Fixing rotate instructions
According to the manual, `ror/rol` instructions has
the following calculation of the number of bit positions:
```
tempCOUNT ← (COUNT & COUNTMASK) MOD SIZE
```
But now, we don't take a `mod size`, while we should.

Also, BIL contains weird assignments like`OF := OF`,
when flag is unchanged, so we will get rid of them.