[build] [WIP] a set of build system optimizations
1. switches build system to compiles setup.ml

2. builds only those plugins that are needed

our build plugin subsystem of the build system is performing lots of
unnecessary work, by rebuilding all plugins every time. The proposed
version builds a plugin only if it is not yet build or if it is older
then its library.

Note, it is a work on progress and contains a hardcoded path to mine
opam.

3. a tool that deletes bap

in case if you messed with the oasis state and can't use uninstall any
more.