introduces progress reports
This commit provides a new feature: progress reports. A new mechanism
that allows plugins to provide a feedback to the system. A rendering
plugin is added that prints this reports in a way that shows the
time profile.

Progress event
--------------

The progress event is a new kind of logging event that allows different
components to provide a feedback about their progress. The supporting
functions like `report_progress` are added to facilitate the usage of
the new event.

Progress reporting plugin
-------------------------

In fact the report plugin is planned to provide more functionality, but
so far it renders progress bars from different components. It requires
ANSI compatible terminal and may clobber output.