moves primus-print on BARE
The `--primus-print-rules` parameter specifies a list of files with
printing rules (though the name is probably bad, as it sounds like the
rules are printed, may be `-by-rules` or `-with-rules` will be better
names).

Each file may contain zero or more BARE rule specifications. If any rule
is not well-formed then a nice (its really nice thanks to the new Sexpar
library) error message is printed and the program is terminated.