changes the default warnings set
OCaml officially disables warning `40` (type disambiguation of record
fields and variant constructors)by default. See ocaml/ocaml#1333. So do
we. This commit will set presets warning to the same set as it is now in
OCaml trunk. Moreover, the same set is used in jbuilder by default, and
the modern OCaml actually started to rely on the absence of the warning
40. For example, new version of Core heavily rely on by using data
constructors that are defined deeply in the internals.

I personally believe that this is a good thing, since the neccessity to
qualify each record or data constructor with its full path usually leads
to lots of open statements that pollute the namespace (unsafe) or to
just to unreadable code. The type disambiguation is a safe way to make
the code much more concise and gives us an opportunity to write more
readable code.

From now on, warning 40 is officially disabled in BAP.

This commit also sync merlin warnings with the compiler warnings. And
let's keep it this way.