[WIP] adds monadic watchers
Previously, watcher were only used for monitoring the event stream and
introspecting, and so far there is a comment that warns users that it
should be only for this (though the comment is superfluous, as the type
system will not allow to affect the machine state in the watcher
handler).

This commit proposes a watcher interface that allows machine
modifications, i.e., the watcher observer will be run the machine
monad. The idea, is that we can subscribe to the observation by name,
from the Lisp machine (using the upcomming rule engine), and generate
observations from observations without any OCaml code. Though the idea
sounds nice, there are several drawbacks. The main is the performance,
as we need to transform each observation into the sexp expression. The
code tries to alleviate this issue by not transforming unwatched
observations, though the core of the problem remains.

Warning: so far, this commit has no dependencies, so it is possible that
it will be discarded or not merged until it is really needed.