[WIP] [UNTESTED] do not mark terms on every block
this may break things and I didn't check that it works right. The idea
is not to update the visited attribute on every block, as it is done
now, but do this once the machine is finished, as every term update
involves the whole rewrite of the program term.

This may have issues as the [finished] stage is heavily overloaded with
different other observers, and it might be the case, that the order of
their invocations matter.

This is a general problem with the [finished] event - as components
may commit their results at this stage, though other components may
collect these results at the same stage, given that there is no ordering
between handlers, the latter might be called before the former. So this
problem should formally addressed.