adds the BAP Rule Engine
BARE is a production system (a forward chaining rule system) is a
pattern matching stream processing rule engine that features non-linear
pattern matching.

Basically, this can be seen as a generalization of the Saluki
solver. Implementation wise, there are two main differences:

1. All facts are represented as a Sexp.t tuples and matching is
performed purely syntactically. In Saluki rules were uninterpreted
functions with arbitrary behavior, though in the paper they were
represented as tuples, except the predicates. So this is closer to
formal behavior of Saluki.

2. Bare operates on the rule level, not on the whole specification
level. In Saluki we were sharing the partial matching state between all
the rules, so if two different rules rely on a the same pattern it will
be matched only once. Though it sounds nice, I still not sure whether
the Saluki solver implemented this stuff correctly. Also, since Bare
doesn't have any a priori information about the input domain, it is
impossible to perform any query optimization, thus we are relying on
Bare users to build efficient rules, based on their local domain
knowledge, and performing this on a rule level, probably, is easier.

Note, though the implementation may look trivial, in comparison with
Rete for example, it is not that naive as it may look like, thanks to
the inherited data sharing of OCaml heap. Although we have just a list
of hypothesis that represent the partial matching state, if you will
look deeper into the representation, then the actual representation, is
that we have a list of patterns, and a list of pairs pointers, where the
first points somewhere in the patterns list, indicating the patterns
that are not yet matched, and the second pointer points to the list of
bindings. Bindings are also shared by different hypotheses that share
the same root hypothesis. So in the memory it is a prefix tree with lots
of data sharing. With all this said, there is a room for optimization,
so far we are performing lots of unnecessary list reordering, though we
can employ some zipper-like structure with O(1) merging.