adds Primus memcheck tool
The memcheck rule is akin to the Valgrind memcheck and detects the same
kind of bugs, i.e., use-after-free, corrupted pointer, memory leaks and
so on.

The tool is memory allocator agnositc and is controlled from the Lisp
machine via two three primitives:

- `(memcheck-acquire ptr len)`
- `(memcheck-release ptr len)`

TODO: make the primitives available to OCaml components

The tool reports all violation via the `memcheck-violate` observation,
that has the following form:

- `(memcheck-violate pos kind)` where `pos` is a program
  location (address) of the violation, and kind is a specific kind of
  violation, e.g., corrupted, use-after-free, def-after-free, and so
  on.

The tool also provides observations of the allocator state, that have
the same name as primitives (probably this is a bad idea to give the
same names to observations and Lisp primitives...):

- `(memcheck-acquire pos (ptr len))` - made every time the
  `memcheck-acquire` primitive is called at the given program position
  `pos`.

- `(memcheck-release pos (ptr len))` - made every time the
  `memcheck-release` primitive is called at the given program position
  `pos`.

Problems:

1. Currently the analysis is broken, due to incorrect comparison of
values, I will fix it soon.

2. Although the tool is allocator agnostic, it doesn't allow checking
several allocators in parallel, as there is no way to distinguish
between different allocators. It would be nice to have this feature, as
it will allow us to handle such bugs as freeing memory using an
incorrect dealloctor.