ignores value id during the comparison
value represents a result of a computation with an attached identifier,
that can be used to attach different properties to the result, using
externalized mapping. Thus the id is not the identity of a value (like
an term identifier that implies that it was a bad idea to name it `id`),
but rather a tag, that allows selection of properties of this particular
result.

This should be addressed carefully in documentation.