[DELETE] looback derived facts into facts
So the problem is that it is now the event processor is not only
generating reports now, but generates facts, that are also processed,
so now it is a fact-processor. Probably it shouldn't be in the print
pluing, that should just print.

Also, the current implementation produces facts derived from the derived
facts before the derived facts.

Another issue - rules currently can communicate with each other as they
do not share the fact stream - thus we need to look at the right hand
side of each rule, and treat them as observation providers. Should we do
this at all, or it would be better to be able to provide only those
facts, that are registered in the system?

I think that this is a bad idea. For example, if the observation
processor create new observations from thin area, we basically create
observations that are not visible from OCaml realm.

I believe, that this commit should be discarded, and instead the
observation processor should be moved in the Lisp machine.

The question is:
How to represent it in the Lisp machine?