[WIP] switches to indexed notation
Instead of trying to use parsexp, we can use sexp that is indexed with
conventional numbers. Such indexing, not only solves a proble with
getting the location information of a parse tree and an AST, but, in
general, allows us to attach arbitrary semantic information to a tree
node. We are using dual-indexing, one index represents identity, that
denotes a node identity, and another index represents equality, such
that all equal nodes have equal equality index.

The main drawback, is that we can't index an arbitrary recursive data
type non-intrusively, as we need to attach the index to each node, thus
we need to define our own data representation for the s-expression,
e.g.,

```
type tree = token indexed
and token = Atom of string | List of tree list
```

The solution above is still partial, as list is a recusive data type on
itself, thus the right solution would be to introduce an indexed list
data type, e.g.,

```
type 'a ilist = 'a node indexed
and 'a node = Nil | Cons of 'a * 'a ilist
```

However, switching from the regular list to the indexed list will
complicate the code base a lot, as we can't use any more the syntactic
sugar for List deconstruction, afaik, OCaml will soon get an ability to
extend the list notation to arbitrary user data type, that will make
this solution viable.

The proposed implementations implements a multipass processing:

1. we take a set of features and parse all source files into a source
repository. The parse tree is an indexed s-expression. The source
repository implements hashconsing and computes equality
relationships. It also records location and file mapping.

2. After all required features are loaded into the source repository,
the repository is parsed in several steps. The first step loads the meta
program, that consists of term-rewriting definitions. The second step
loads the definitions while applying all term-rewriting rules. The
second step can be seen as a symbolic computation step.