[WIP] separates parsing steps
The steps should be applied on the per source repository basis, not on
per module basis. Now, when we are parsing definitions the meta program
is totally ready and we have all meta definitions. Since we need to
apply macro definitions during the parsing we need all of them. Since we
do not apply regular lambda applications at this stage, the order of
parsing is not relevant.

We may later add a pass that will resolve calls, so far, we will use the
late binding.