adds symbols to Primus Lisp
Adds a new term kind - symbol. A symbol is a first class value, that
has the same runtime representation as a word (we use interned strings
as underlying representation). A symbol is introduced in a program
using a regular `'<sym>` literal, e.g., `'cat`.

This commit also reverts character literals back to `?<char>` Emacs
Lispish notation (instead of '<char>' that was used for some time in
this branch).