changed internal foreach interface
In PowerPC arch manual all bits, bytes etc.
are indexed starting from the most significant part.
And a design concept in out lifter is to let user
to see the same approach on his level, and to index
everything as it should be done in our internal code.
So basicaly, in DSL everything works like user expect to,
in RTL implementation - as we expect to. The only exception
is foreach, that breaks this concept, and use weird
index approach. So this PR solves this issue. And now
index numeration starting from modt significant bit exists
only in DSL