changes the merging strategy of information sources (#759)
In the previous implementation, if we had several sources of information then we wait until all are ready, and then join them. The problem was that some source might not be ready, as they depend on some signal, that either doesn't happen at all, or will happen after the information is produces (the deadlock). The new implementation will output information as soon, as it is available, while still merging it. 

This commit has a potential of breaking the symbolization process, i.e., it may change how symbols are assigned to program locations. Though during the review we managed to persuade ourself that the order is preserved at least with the given setup, there still some room for pessimism.