increases precision of the taint analysis
if a reference already points to the tainted object we overwrite it
with the new taint, while previously we were merging the old set of
referenced tainted objects with the new set.

It is still not precise, as it doesn't account for the assignments
that are not tainted.

Also fixes a couple of bugs, and introduces a nicer mechanism for
generalization tainting functions (i.e., selectors).