improves performance of taint propagation
First of all, the previous note on the taint propagation precision was
wrong, it is precise, as when we rewrite a reference with an untainted
object we are associating it with an empty set of tainted objects. To
minimize the number of references that we track, we can just remove a
references instead of associating it with an empty set. This is
effectively what this commit is doing.