Fixed a search of a block destination
In the reconstructor, a destination of a block obtained
directly from bil, that doesn't correct, as bil doesn't
always contain jumps to proper places, e.g. when a relocation
should be applied to a lifted instruction. And if the relocatable
brancher is used, then cfg contains an edges to a proper
destinations, but this information is ignored during calls lookup
in reconstructor.

This PR solves this problem by searching a destination
directly from cfg, obtained from previous stages. So all
edges that lead to a subroutines starts are taken in account.