fixes the search of destinations in the reconstructor (#761)
fixes the search of destinations in the reconstructor

In the reconstructor block destinations are computed directly from the BIL representation, that is not correct, as BIL doesn't take into account the relocations. And if the relocatable brancher is used, then the CFG contains correct edges, but this information is ignored when we build a set of call destinations in the reconstructor. This PR addresses this problem by searching a destination directly in CFG.