removes assertions about taint object correctess
We can't actually enforce by the abstract interface the invariant of
the taint object, that it must have an associated kind, as we still
expose the `of_value` function, that injects any value into the object
type. So it is possible to associate with a value anything besides the
taint object. Thus instead of doing the assertion false, we assume
this and silently ignore all other values. Honestly, I don't like the
silent part of this approach, so in future it may change.