fixed operands indexes
Got an exception in CLO and CLZ instructions, array out of bounds.
So  the reason is that llvm claims that there are only two operands:
```
$: echo "clo \$3, \$4" | llvm-mc-3.8 --arch=mips --show-inst-operands --show-inst --show-encoding

clo	$3, $4                  # encoding: [0x70,0x83,0x18,0x21]
                                # <MCInst #590 CLO
                                #  <MCOperand Reg:322>
                                #  <MCOperand Reg:22>>

```
So my previous suggesting (to prefix unused operands) was not correct.
Also I added a better check of the operand array bounds - so if
something is wrong, we will see a better message:
```
$: echo "70 83 18 20" | bap-mc --show-bil --arch=mips --show-insn=asm
clz $3, $4
{
  special (Lifter: instruction CLZ doesn't have an operand with index 2)
}
```