rectifies the interrupt handling
Previously a interrupt was halting a machine with an exception
without any option to handle it. The new implementation provides an
observation when an interrupt happens and, by default, proceeds to
the interrupt fall-through destination. It is possible now, to attach
arbitrary behavior to the interrupt using the observation mechanism,
or by providing a method for the `interrupt` signal in the Lisp
interpreter, e.g.,

```lisp
(defmethod interrupt (n)
   (when (= AH 8)
     (msg "DOS print_string is called")))

```