deprecates primus-propagate-taint in favor of primus-taint
The new interface makes much more sense with Primus and is much easier
to use, as there is no need to explicitly enable anything. The
propagation is enabled by default. It is possible to choose a default
policy out of provided two policies or implement a custom policy. The
policy is selectable on the per taint kind basis, thus at the same time we
can propagate different taints with different policies.

I still didn't expose all the functions to the lisp. And didn't expose
the `taint-policy-select` function, as well as the
`taint-policy-set-default`. This would be done in the next commit, I hope.