revives the primus-propagate-taint plugin
We shouldn't really deprecated it until we deprecate the old taint
propagation pipeline (and I'm not sure whether we will do this, or
need to do this). The (new) purpose of the primus-propagate-taint
plugin is to enable integration between the new Taint Analysis
Framework and the old taint propagation pipeline, i.e., it establishes
mapping between the taint attributes attached to program term, and
taints attached to computations.