adds the powerpc lifter (#739)
This is a squashed commit, the full change log is below

* powerpc lifter

It's a initial PR in order to bring PowerPC lifter
into a life. It contains only one group of instructions,
just for discussing, along with a service modules like
`Powerpc_model`, `Powerpc_dsl` and others.

* PowerPC lifter is ready for review

* slighly changed documentation

* changed internal foreach interface

In PowerPC arch manual all bits, bytes etc.
are indexed starting from the most significant part.
And a design concept in out lifter is to let user
to see the same approach on his level, and to index
everything as it should be done in our internal code.
So basicaly, in DSL everything works like user expect to,
in RTL implementation - as we expect to. The only exception
is foreach, that breaks this concept, and use weird
index approach. So this PR solves this issue. And now
index numeration starting from modt significant bit exists
only in DSL

* minor

* doc edited

* minor

* added PowerPC to bapdoc

* fixed few bugs in store and load insns

llvm decoded immediate values for us in those instructions, so we must
not to do it ourselfs

* edited opam file

* updated bap opam file

fixed path to new signatures

* renamed fields in cpu model

* mv all PowerPC files into single plugin

* files renaming

* removed bap-powerpc from bap-doc

* refactoring in progress

changes:
1) removed all signed operators.
   so a signed version of comparison/division/modulo depends totaly
   from operands
2) removed logical shifts (lsl and lsr). Now shift is logical if
   operands are unsigned and arithmetical if signed.
3) refactored branch and add instructions

* modularization: div, sub, mul + bugs fixed

* adds using Map.find instead of String.Map.find

* fix documentation

* added init function to every instruction module

* fixed example in docs