fixes mips release 6 problems
As for now, we have a problem with disassembling and lifting code built
for Mips VI release. Although `llvm-mc` works perfectly, if we pass a
special argument, e.g. `-mattr=mips32r6`.

One possible solution would be to add additional arches for 6th release,
like we did it for ARM architecture. Like "mips32r6".
But at first it's not very beautiful, because we will need to add 4(!!!)
architectures: for 32 and 64 bit, for little and big endian. At second,
it may hit as in future with new Mips release, whenever it will be.
At third, on the mips lifter side, this addition will face us with a
bunch of new problems, because we will have to handle a bigger ammount
of architectures.
And at fourth, the most important one. In ARM case, llvm knows how to
handle different arm types, there is a special modules/functions for
this purpose. So, `armv7` is a valid architecture name.
But it is not true in Mips case.

A better solution would be add a command line argument for this purpose.

Let's look how `llvm-mc` creates subtarget:
```
TheTarget->createMCSubtargetInfo(TripleName, MCPU, FeaturesStr));
```
where `FeaturesSet` is exactly the place where "-mattr=mips32r6" goes.

And how we do the same:
```
target->createMCSubtargetInfo(triple, cpu, "");
```
So next question is: How to transfer features from cmdline ?
But it turned out, that in `llvm-mc`, the `mips32r6` could transfered
through `-mcpu` cmdline argument. At least, llvm tests use it in this
way. And what I suggest is just to add the same cmdline argument.
Anyway, it already exists in disassembler interface, but not for
all functions. So we need just fix it and also add it in `Bap_project`
interface.