adds more options to the primus-limiter
Now it is possible to set the maximum length of a trace (path) in
blocks, instructions, terms, and expressions. The kind of a counter
that should be used is defined via the suffix, where `b` stands for
blocks, `i` for instrucitons, and so on, e.g., `1000b` means a
thousand of basic blocks.

By default, if no prefix is set, we use expressions. Since the Lisp
Interpreter, as well as any other linked code, is not required to
express its computation in terms of blocks, it is better to use
expressions, since the Lisp machine expresses its computation on this
level.