bumps the bap version to 1.5.0-dev (#790)
* bumps the bap version to 1.5.0-dev

to prevent the confusion with the already released version

* removes signature update from the test script

The signatures are already installed in previous stages, so they are
not needed. Also the development version doesn't have released
artifacts, so by default the update won't work unless an explicit url
is specified. So it is easier just to remove it.