fixes BIL normalization procedure (#789)
* intentation changes

apparently ocp-indent was broken, when I was editingt this file the
last time.

* adds hoisting non-generatives from specific points

This commit extends BNF1 with an extra condition, that no load or
store expressions in the following positions:

 1. the right-hand side of the let expression;
 2. address or value subexpressions of the store expression;
 3. storage or address subexpressions of the load expression.

Those positions are the specific positions where the BNF2
normalization will involve rearragement and copying of
expressions. Thus we need to ensure that in those positions all
expressions are generative (i.e., do not produce observable (by other
expressions) effects). We enforce this restriction by hoisting all
non-generative expressions into statements. Since this transformation
produces new statements its impossible to apply it on the exression
level, so we are putting BNF1 as a precondition for the
`Exp.normalize` function.