fixes relative addresses in llvm backend (#794)
* fixes relative addresses in llvm backend

There was a kind of bug in llvm backend: relative address
for any binary entry was defined as non-negative value.
Probably, it wasn't a very bad choice, since we don't want to
deal with negative addresses. But from other point of view we
always convert relative address to absolute and negative
address should not happen. And we actually inject wrong
information in ogre when store non-negative numbers:
converting them to absoulte addresses assigns a real value to
every entry, dependless if it had meaningful address or not at all.

* fixed types for relative addresses