provides jumping observation and resolve_xxx functions
The jumping observation is made just before a jump is taken and has
two parameters: jump condition and jump destination. This is a much
more precise and useful observation than the `{enter,leave}-term`
one, as it doesn't loose the origin of values that define the
jump (i.e., we can track (taint) both the condition and the
destination).

We also added three functions to the Linker interface, mostly for
convenience. These functions resolve code names into addresses,
symbolic names, and tid.