rectifies the primus-lisp plugin
So far the plugin loads some stuff during the link time, some after
the configuration is available, and the rest during the analysis phase
when it is invoked as a pass. It may even be invoked  twice.

Now it is rectified. Everything that doesn't require an access to the
project data structure is loaded automatically just after the
configuration phase has finished. The rest (e.g., linking with the
program) is done during the analysis.