more efficient Lisp stack and a bug fix in signal handlers
1. Provides an efficient stack that now is equipped with a map data
structure that allows us effeciently query for locally bound
variables.

In Primus Lisp variables live in two scope:
 1. the Primus global such as CPU registers, parameters, etc.
 2. the Primus Lisp local scope that is populated by lexically scoped
    local variables such as function parameters and let bound
    variables.

Any Lisp operation on a variable involves checking whether this is a
local variable or global variable. Since we're using lexical scoping
we should be able to distinguish between them in compile time. And
that's what we will eventually do in the future. But so far it
requires to many changes to the interpreter. Previously, we were using
an assoc list, that was traversed every time we read or set a
variable. The new implementation uses a balanced tree that maps
variables to the total number of their occurences on the stack. That
makes the check to be logarithmic in the number of unique variable
names on the stack.

Few other small optimizations:

a) faster function frame costructor, that is not monadic (as it
shouldn't be), and that is not preserving the order of arguments (so
that we can append them in the reversed order, using fold left), and
that counts the total number of arguments, so that we don't call
List.length when we're poping the frame off.

b) a faster frame pushing mechanism - instead of just appending
arguments, we are prepending them in the reverse order. This is not
only faster, but since the frame is in the reverse order already, it
preserves the existing order (so no changes in semantics)

2. Fixes an nasty bug in the signal dispatcher. When a signal is
dispatched to several methods each method ignores effects from other
methods as it resets the state as it was before the dispatcher was
entered.