keeps primitives during linking and moves init_env to loader
1. we do not want primitives, that were registered by other components
to be dropped when a new program is linked. In fact, primitives is a
complete different beast than other definitions, as they are defined
by OCaml modules (i.e., plugins) rather than by the Lisp code, so
probably they shouldn't be a part of the program data structure at
all. Though having them in the program makes things much easier.

Besides, when we drop primitives during the program linking time, we
are discarding all primitives that were registered prior to the call
to program_link. So far, just by accident, the link_program function
was called before any other components (since it was registered last -
after all other lisp components). This is so fragile, that I consider
this is a bug. Nasty one.

2. The `init_env` function in the Primus.Lisp module was adding
special variables of the form `%<id>` and `@<id>` that hold addresses
of the correspondingly named program terms. It was using the local
variables stack. Neither should it be using the local stack for
storing this variables (as they are globals by their nature), nor this
should be done during the linking time of the program. This function
is now moved to the loader (that is the component that's responsible
for setting up the environemnt).